requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
nltk==3.8.1
pypdf==4.2.0
python-docx==1.1.0

sentence-transformers==2.6.1
transformers==4.39.3
torch==2.2.2

langchain==0.2.6
langchain-core==0.2.12
langchain-community==0.2.6
langchain-groq==0.1.3
langchain-chroma==0.1.1

chromadb==0.5.3
